requirements.txt
matplotlib>=3.5.0
numpy>=1.20.0
pandas>=1.3.0
seaborn>=0.11.0
plotly>=5.3.0
pyyaml>=6.0